Implement zooming of slide, closes part 1 of #66
Inside scribbler, as we put this button with the set of
scribbling/highlighting buttons.

Save zoom level as float, and offset as relative position in slide. This
is so that we can use it with different slide sizes. The shift would be a
fixed number of pixels if we used a matrix. The zoom is just a cairo
context transform, without re-rendering the image, so if zoomed too much
the slide will look pixelated.

Refactor all click-to-position-in-slide computations to a single
function that takes into account the zoom level etc. Also update the
function in ui that returns the in-page position to take this into
account. The difference here is that in notes mode, the slide is the
half of the page that is shown by the widget that is clicked.

Add 2 buttons:
- one to zoom in, which starts drawing a zoom-preview rectangle. Perform
  the zoom as the button is released. TODO: visual feedback that button
  was pressed, before we click on the slide for zooming?

- one to go back to no zoom (also performed on page change). It is
  disabled by default and is enabled when zooming is in effect.